feat(lang): add Typescript support
refactor(lang): wrap Rust query-logic in a struct.

Using the rust implementation as a base turned out to be really hard
because the rust implementation was really hard to understand, so a
big refactoring for code simplification also happened to the Rust
version of the library.

fix!(rust): Use the instrumented struct as part of the function, not module

BREAKING: in Rust, the name of the structure for instrumented methods is
now included in the `function` field, not the `module` anymore.

Golang refactoring will happen later, but this makes changing code a lot easier.